+++ ml_project/requirements.txt
@@ -1,6 +1,6 @@
 antlr4-python3-runtime==4.8
 attrs==21.2.0
-Bottleneck==1.3.2
+Bottleneck==1.3.3
 certifi==2020.12.5
 chardet==4.0.0
 confuse==1.4.0
@@ -13,7 +13,7 @@ idna==2.10
 ImageHash==4.2.0
 iniconfig==1.1.1
 Jinja2==2.11.3
-joblib==1.0.1
+joblib==1.1.0
 kiwisolver==1.3.1
 MarkupSafe==1.1.1
 matplotlib==3.4.2
@@ -41,11 +41,11 @@ scikit-learn==0.24.2
 scipy==1.6.3
 seaborn==0.11.1
 six==1.16.0
-tangled-up-in-unicode==0.0.7
+tangled-up-in-unicode==0.2.0
 text-unidecode==1.3
 threadpoolctl==2.1.0
 toml==0.10.2
 tqdm==4.60.0
 typing-extensions==3.10.0.0
 urllib3==1.26.4
-visions==0.7.1
+visions==0.7.5
