+++ ml_project/setup.py
@@ -1,9 +1,20 @@
 from setuptools import find_packages, setup
 
-from ml_project.utils.read_files import parse_requirements
-
 
 REQUIREMENTS_PATH = 'requirements.txt'
+
+
+def parse_requirements(requirements_path: str) -> list[str]:
+    """Parses pip requirements file
+    :param requirements_path: path to requirements file
+    :return: list of required packages
+    """
+    with open(requirements_path) as f:
+        requirements_file_lines = f.read().splitlines()
+    processed_lines = map(lambda x: x.strip(), requirements_file_lines)
+    required_packages = filter(lambda x: x and not x.startswith('#'), processed_lines)
+    return list(required_packages)
+
 
 required_packages = parse_requirements(REQUIREMENTS_PATH)
